Observe fxa auth state and notify menu builder in fenix account manager
