+++ app/src/main/java/org/mozilla/fenix/home/HomeToolbarMenu.kt
@@ -24,6 +24,7 @@ import mozilla.components.concept.sync.OAuthAccount
 import mozilla.components.support.ktx.android.content.getColorFromAttr
 import org.mozilla.fenix.FeatureFlags
 import org.mozilla.fenix.R
+import org.mozilla.fenix.components.accounts.FenixAccountManager
 import org.mozilla.fenix.experiments.ExperimentBranch
 import org.mozilla.fenix.experiments.Experiments
 import org.mozilla.fenix.ext.components
@@ -266,6 +267,8 @@ class HomeToolbarMenu(
     }
 
     init {
+        val accountManager = FenixAccountManager(context, lifecycleOwner)
+
         val menuItems = if (FeatureFlags.toolbarMenuFeature) {
             newCoreMenuItems()
         } else {
@@ -283,43 +286,11 @@ class HomeToolbarMenu(
         }
 
         // Observe account state changes, and update menu item builder with a new set of items.
-        context.components.backgroundServices.accountManagerAvailableQueue.runIfReadyOrQueue {
-            // This task isn't relevant if our parent fragment isn't around anymore.
-            if (lifecycleOwner.lifecycle.currentState == Lifecycle.State.DESTROYED) {
-                return@runIfReadyOrQueue
-            }
-            context.components.backgroundServices.accountManager.register(object : AccountObserver {
-                override fun onAuthenticationProblems() {
-                    lifecycleOwner.lifecycleScope.launch(Dispatchers.Main) {
-                        onMenuBuilderChanged(
-                            BrowserMenuBuilder(
-                                menuItemsWithReconnectItem
-                            )
-                        )
-                    }
-                }
-
-                override fun onAuthenticated(account: OAuthAccount, authType: AuthType) {
-                    lifecycleOwner.lifecycleScope.launch(Dispatchers.Main) {
-                        onMenuBuilderChanged(
-                            BrowserMenuBuilder(
-                                menuItems
-                            )
-                        )
-                    }
-                }
-
-                override fun onLoggedOut() {
-                    lifecycleOwner.lifecycleScope.launch(Dispatchers.Main) {
-                        onMenuBuilderChanged(
-                            BrowserMenuBuilder(
-                                menuItems
-                            )
-                        )
-                    }
-                }
-            }, lifecycleOwner)
-        }
+        accountManager.observeAccountState(
+            menuItemsWithReconnectItem,
+            menuItems,
+            onMenuBuilderChanged
+        )
     }
 }
